Split CaseReviewTab into multiple files
